server_side/trdi_python_requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.5
cffi==1.11.5
Click==7.0
cryptography==2.4.2
enum34==1.1.6
Flask==1.0.2
future==0.17.1
idna==2.8
ipaddress==1.0.22
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10
libusb==1.0.22b2
MarkupSafe==1.1.0
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyusb==1.0.2
PyYAML==3.13
qrcode==6.0
serial==0.0.86
six==1.12.0
Werkzeug==0.14.1
